requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
tenacity>=8.2.3
pywinrm>=0.4.3
email-validator
